requirements.txt
matplotlib
google-generativeai
